fix: refactor io-uring to drive the queue to completion (#1498)
Previously, we were encountering issues with the io-uring writes where
we discovered we were not completing writes.

After extensively studying `liburing` and the `io-uring` crate, I
realized our submission/processing loop was unnecessarily complicated
and did not continuously drive the queue until completion. The crux of
the fix is the `backlog` dequeue that keeps track of the backlogged
entries and the `outstanding` hashmap that keeps track of the entries
we've submitted to the kernel. CQEs that don't indicate a complete write
are re-queued for submission.

Closes: #1293 
Closes: #1488

## References

-
<https://github.com/tokio-rs/io-uring/blob/f297f5bdb4802915b48a532edf9b817d6c91de8f/examples/tcp_echo.rs>
-
<https://github.com/axboe/liburing/blob/e1ef1e680ee38ed9116989155fca47921698c25f/examples/io_uring-cp.c>