perf: avoid extra copying when comparing `HashOrRlp` (#1505)
The change I previously made to `HashOrRlp` to reduce some code
redundancy added extra overhead in the form of unnecessary copying when
comparing two `HashOrRlp` instances for equality or hashing them. That
added a noticable increase to the average propose+commit benchmark
times.

This change reverts to the prior method of comparing without copying.

I have also moved the `HashType` into its own module with `TrieHash` and
`HashOrRlp` as submodules.

Closes: #1502