fix(range-proofs): generate proofs for lower and upper bounds correctly
When we generate range proofs, we must prove the beginning of the
requested range even if the requested start key is not present in the
trie. This is necessary so that clients can prove there is no data
between the proven start and the first yielded key. Similarly, we must
prove the requested end key so that the client can prove there is no
data between the last yielded key and request key. If (and only if) we
truncate the set of key-values returned should we prove a different key
than the one requested. Additionally, we must ensure all keys yielded
are within the requested bounds.

Before this change, we were only generating proofs for the first key
found. This makes it impossible for clients to verify that there is no
data between the requested start key and the first key provided.
Similarly, were also always generating proofs for the last key found. If
the client did not request an upper bound, it's superfluous information
as hashing the key-value set with only the lower bound must still yield
a root node with the expected hash. If the client did request an upper
bound and we didn't truncate the range, it makes it impossible for
clients to prove there is no data between the last key provided and the
last key requested.