test: enhance aws-launch with show and terminate options
Add three new options to aws-launch.sh for managing EC2 instances:

--show: Display currently running instances in a region
  Shows instance details including ID, launch time, IP, type, and name tag.
  Filters by key-name 'rkuris' and running state.
  Example: ./aws-launch.sh --show --region us-west-2

--terminate-mine: Terminate all instances created by current user
  Identifies instances by Name tag prefix matching $USER.
  Provides feedback on instances being terminated.
  Example: ./aws-launch.sh --terminate-mine --region us-east-1

--terminate: Terminate specific instances by ID
  Accepts one or more instance IDs as arguments.
  Example: ./aws-launch.sh --terminate i-1234567890abcdef0 i-0987654321fedcba0

All three options can be combined with --region to target specific regions.
They cannot be used with other launch-related options to prevent confusion.

Also added "--end-block" to override the number of blocks. This is
used to get a firewood database at a specific height, and was useful
during debugging.