chore: reorganize hash helper
This is a continuation of #1137; which, had it been merged, would have
fixed #1302 long before #1557.

This resolves the extensive merge conflicts from the original and
continues the refactor of the hash helper functions simplifying their
interface. A bit of early progress toward making ethhash a runtime option
was also made by migrating the `cfg(...)` usage deeper into the helper.

Co-authored-by: Suyan Qu <36519575+qusuyan@users.noreply.github.com>